+++ smif/app/src/components/ConfigForm/General/PropertySelector.js
