+++ smif/app/src/components/ConfigForm/ProjectOverview/ProjectOverviewItem.js
