+++ smif/app/src/containers/ConfigForm/NarrativeConfig.js
@@ -4,12 +4,12 @@ import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import { Link, Router } from 'react-router-dom'
 
-import { fetchNarrative } from '../actions/actions.js'
-import { fetchNarrativeSets } from '../actions/actions.js'
+import { fetchNarrative } from '../../actions/actions.js'
+import { fetchNarrativeSets } from '../../actions/actions.js'
 
-import { saveNarrative } from '../actions/actions.js'
+import { saveNarrative } from '../../actions/actions.js'
 
-import NarrativeConfigForm from '../components/NarrativeConfigForm.js'
+import NarrativeConfigForm from '../../components/ConfigForm/NarrativeConfigForm.js'
 
 class NarrativeConfig extends Component {
     constructor(props) {
