+++ smif/app/src/containers/ConfigForm/SosModelRunConfig.js
@@ -4,14 +4,14 @@ import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import { Link, Router } from 'react-router-dom'
 
-import { fetchSosModelRun } from '../actions/actions.js'
-import { fetchSosModels } from '../actions/actions.js'
-import { fetchScenarios } from '../actions/actions.js'
-import { fetchNarratives } from '../actions/actions.js'
+import { fetchSosModelRun } from '../../actions/actions.js'
+import { fetchSosModels } from '../../actions/actions.js'
+import { fetchScenarios } from '../../actions/actions.js'
+import { fetchNarratives } from '../../actions/actions.js'
 
-import { saveSosModelRun } from '../actions/actions.js'
+import { saveSosModelRun } from '../../actions/actions.js'
 
-import SosModelRunConfigForm from '../components/SosModelRunConfigForm.js'
+import SosModelRunConfigForm from '../../components/ConfigForm/SosModelRunConfigForm.js'
 
 class SosModelRunConfig extends Component {
     constructor(props) {
